Import OCCT 6.7.1
 Remove unneeded folders, they are very large:
   data samples doc dox

 # Remove useless WOK folders
 rm -rf drv/*/.adm

 # Fix permissions (on Unix):
 find * -type f -exec chmod a-x {} \;

 # Autotools scripts
 chmod a+x build_configure

 # Windows scripts
 find adm/msvc  -type f -exec chmod a+x {} \;

 # Shell and bat scripts
 chmod a+x \
    adm/templates/draw.sh  \
    adm/templates/draw.bat  \
    adm/templates/env.sh.in  \
    adm/templates/env.bat.in  \
    adm/templates/sample.bat  \
    codeblocks.sh  \
    custom.bat  \
    custom.sh.in  \
    draw.bat  \
    draw.sh  \
    env.bat  \
    env.sh  \
    gendoc.bat  \
    gendoc  \
    msvc.bat  \
    xcode.sh

+++ src/RWStl/RWStl.cxx
@@ -5,8 +5,8 @@
 //
 // This file is part of Open CASCADE Technology software library.
 //
-// This library is free software; you can redistribute it and / or modify it
-// under the terms of the GNU Lesser General Public version 2.1 as published
+// This library is free software; you can redistribute it and/or modify it under
+// the terms of the GNU Lesser General Public License version 2.1 as published
 // by the Free Software Foundation, with special exception defined in the file
 // OCCT_LGPL_EXCEPTION.txt. Consult the file LICENSE_LGPL_21.txt included in OCCT
 // distribution for complete text of the license and disclaimer of any warranty.
@@ -30,13 +30,13 @@
 #include <stdio.h>
 #include <gp_Vec.hxx>
 
-
 // constants
-static const int HEADER_SIZE           =  84;
-static const int SIZEOF_STL_FACET      =  50;
-static const int STL_MIN_FILE_SIZE     = 284;
-static const int ASCII_LINES_PER_FACET =   7;
-static const int IND_THRESHOLD         = 1000; // increment the indicator every 1k triangles
+static const size_t HEADER_SIZE           =  84;
+static const size_t SIZEOF_STL_FACET      =  50;
+static const size_t STL_MIN_FILE_SIZE     = 284;
+static const size_t ASCII_LINES_PER_FACET =   7;
+
+static const int IND_THRESHOLD = 1000; // increment the indicator every 1k triangles
 
 //=======================================================================
 //function : WriteInteger
